feat: add access controll to task resource

+++ apps/api/src/app/tasks/tasks.service.ts
@@ -4,27 +4,49 @@ import { UpdateTaskDto } from './dto/update-task.dto';
 import { Repository } from 'typeorm';
 import { Task } from './entities/task.entity';
 import { InjectRepository } from '@nestjs/typeorm';
+import { OrganizationPermissionService } from '../organizations/organization-permissions.service';
+import { OrganizationsService } from '../organizations/organizations.service';
+import { PermissionsEnum } from '../rbac/permission';
 
 @Injectable()
 export class TasksService {
 
   constructor(
     @InjectRepository(Task)
-    private readonly taskRepository: Repository<Task>
+    private readonly taskRepository: Repository<Task>,
+    private readonly permissionService: OrganizationPermissionService,
+    private readonly organizationService: OrganizationsService,
   ) {}
 
-  create(createTaskDto: CreateTaskDto): Promise<Task> {
+  async create(userId: number, organizationId: number, createTaskDto: CreateTaskDto): Promise<Task> {
+    const org = await this.organizationService.findOne(organizationId, userId);
+    await this.permissionService.checkOrganizationPermission(org.id, userId, PermissionsEnum.TASK_CREATE);
     const task = new Task();
     task.title = createTaskDto.title;
     task.description = createTaskDto.description;
+    task.organization = org;
     return this.taskRepository.save(task)
   }
 
-  findAll(): Promise<Task[]> {
-    return this.taskRepository.find();
+  async findAll(userId: number, organizationId: number): Promise<Task[]> {
+    const org = await this.organizationService.findOne(organizationId, userId);
+    await this.permissionService.checkOrganizationPermission(
+      org.id, userId, PermissionsEnum.TASK_VIEW
+    );
+    return this.taskRepository.find({
+      where: {
+        organization: {
+          id: organizationId,
+        },
+      },
+    });
   }
 
-  async findOne(id: number): Promise<Task> {
+  async findOne(userId: number, organizationId: number, id: number): Promise<Task> {
+    const org = await this.organizationService.findOne(organizationId, userId);
+    await this.permissionService.checkOrganizationPermission(
+      org.id, userId, PermissionsEnum.TASK_VIEW
+    );
     const task = await this.taskRepository.findOneBy({ id });
     if (!task) {
       throw new NotFoundException(`Task id ${id} not found`);
@@ -32,16 +54,22 @@ export class TasksService {
     return task;
   }
 
-  async update(id: number, updateTaskDto: UpdateTaskDto) {
-    const task = await this.findOne(id);
+  async update(userId: number, organizationId: number, id: number, updateTaskDto: UpdateTaskDto) {
+    const task = await this.findOne(userId, organizationId, id);
+    await this.permissionService.checkOrganizationPermission(
+      organizationId, userId, PermissionsEnum.TASK_EDIT
+    );
     if (updateTaskDto.status) task.status = updateTaskDto.status;
     if (updateTaskDto.title != null) task.title = updateTaskDto.title;
     if (updateTaskDto.description != null) task.description = updateTaskDto.description;
     return this.taskRepository.save(task);
   }
 
-  async remove(id: number): Promise<void> {
-    const task = await this.findOne(id);
+  async remove(userId: number, organizationId: number, id: number): Promise<void> {
+    const task = await this.findOne(userId, organizationId, id);
+    await this.permissionService.checkOrganizationPermission(
+      organizationId, userId, PermissionsEnum.TASK_DELETE
+    );
     await this.taskRepository.remove(task);
   }
 }
